+++ packages/catalog/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed the standard Xiaomi provider default model to `mimo-v2.5`, matching the upstream-supported MiMo model. ([#4062](https://github.com/can1357/oh-my-pi/issues/4062))
+
 ## [16.2.12] - 2026-07-01
 
 ### Breaking Changes